Remove exportertest pkg
