+++ sdk/metric/processor/processortest/test.go
@@ -15,6 +15,7 @@
 package processortest
 
 import (
+	"context"
 	"fmt"
 	"strings"
 	"sync"
@@ -47,10 +48,6 @@ type (
 	}
 
 	// Output implements export.CheckpointSet.
-	//
-	// TODO(#872) make this internal and update callers to use the
-	// Processor, single Checkpointer, or Exporter facilities in
-	// this package.
 	Output struct {
 		m            map[mapKey]mapValue
 		labelEncoder label.Encoder
@@ -72,6 +69,15 @@ type (
 	Processor struct {
 		export.AggregatorSelector
 		output *Output
+	}
+
+	// Exporter is a testing implementation of export.Exporter that
+	// assembles its results as a map[string]float64.
+	Exporter struct {
+		export.ExportKindSelector
+		output      *Output
+		ExportCount int
+		InjectErr   func(export.Record) error
 	}
 )
 
@@ -129,9 +135,6 @@ func (c *testSingleCheckpointer) CheckpointSet() export.CheckpointSet {
 // (from an Accumulator) or an expected set of Records (from a
 // Processor).  If testing with an Accumulator, it may be simpler to
 // use the test Processor in this package.
-//
-// TODO(#872): This class should be made internal, and callers should either
-// use a test Processor or a test Exporter to use these facilities.
 func NewOutput(labelEncoder label.Encoder) *Output {
 	return &Output{
 		m:            make(map[mapKey]mapValue),
@@ -209,8 +212,6 @@ func (o *Output) ForEach(_ export.ExportKindSelector, ff func(export.Record) err
 // to a map for use in testing.  The value taken from the record is
 // either the Sum() or the LastValue() of its Aggregation(), whichever
 // is defined.  Record timestamps are ignored.
-//
-// TODO: Move this into an exporter test helper package.
 func (o *Output) AddRecord(rec export.Record) error {
 	key := mapKey{
 		desc:     rec.Descriptor(),
@@ -279,4 +280,45 @@ func (o *Output) AddAccumulation(acc export.Accumulation) error {
 			time.Time{},
 		),
 	)
+}
+
+// NewExporter returns a new testing Exporter implementation.
+// Verify exporter outputs using Values(), e.g.,:
+//
+//     require.EqualValues(t, map[string]float64{
+//         "counter.sum/A=1,B=2/R=V": 100,
+//     }, exporter.Values())
+//
+// Where in the example A=1,B=2 is the encoded labels and R=V is the
+// encoded resource value.
+func NewExporter(selector export.ExportKindSelector, encoder label.Encoder) *Exporter {
+	return &Exporter{
+		ExportKindSelector: selector,
+		output:             NewOutput(encoder),
+	}
+}
+
+func (e *Exporter) Export(_ context.Context, ckpt export.CheckpointSet) error {
+	e.ExportCount++
+	return ckpt.ForEach(e.ExportKindSelector, func(r export.Record) error {
+		if e.InjectErr != nil {
+			if err := e.InjectErr(r); err != nil {
+				return err
+			}
+		}
+		return e.output.AddRecord(r)
+	})
+}
+
+// Values returns the mapping from label set to point values for the
+// accumulations that were processed.  Point values are chosen as
+// either the Sum or the LastValue, whichever is implemented.  (All
+// the built-in Aggregators implement one of these interfaces.)
+func (e *Exporter) Values() map[string]float64 {
+	return e.output.Map()
+}
+
+func (e *Exporter) Reset() {
+	e.output.Reset()
+	e.ExportCount = 0
 }
